Transition to templated dynamic and static object and function types.
This change sets the stage for programs with fully known memory allocation, providing `mallocStruct`s for addition parameters that need to be allocated during function allocation.

The front-facing changes are minimal, since appropriate functions are chosen based on the SFINAE principle using `std::enable_if`, based on an unambiguous assessment on whether or not a system's dimensions indicate if it is statically initialized at compile-time, or if it is dynamic (it is considered dynamic if any of its elements is dynamic, see `aux/CompileTimeQualifiers.hpp`). For all systems, default parameters refer to the dynamic specialization of that object.