Add ASCII graphs using graph-easy. Add stub for a basic CLI. Add preliminary interface for encryption (ED22519).
The encryption interface consists of a signature field, in which 12 64-bit ints are exposed for the public key (4*8 bytes), and the message signature (8*8 bytes). This allows a program to be run with a set of trusted public keys, from which input is accepted. In turn, the program should have its own public key that is known by the user (either randomly defined, or user-specified). This public key can also be used to check message integrity when paired with the signature.